feat: Add Siamese_Model for feature extraction and classification
This commit adds a new file, Siamese_Model.py, which contains the implementation of a Siamese neural network (SNN) for feature extraction and classification. The SNN takes two input images and extracts features using a pretrained model. It then concatenates the features and passes them through a classifier to predict a binary output. The last layer of the classifier has been modified to fit the specific problem.

The SNN is implemented using PyTorch and includes various layers such as linear, ReLU, sigmoid, and dropout. The changes also include updating the last layer of the efficientNet_b0 model to fit the problem, as commented out in the code.

This feature will improve the accuracy of the classification model and enable better analysis of the input images.